requirements.txt
pillow
requests_ratelimit_adapter
requests_cache
requests